requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-storage==2.13.0
flask-cors==4.0.0
